Add the command_intent to execute commands onto tiles
This allows the player to delete a tile.

There was initially a wierd situation where if a tile was deleted, it would screw up the gameboard.
The gameboard does not keep track of rows and columns, there has to be a tile in each position
for it to render correctly. So now I just set a piece of the board as 'nothing' rather than delete it
from the database.

I also uncovered a bug where moving onto a deleted tile ended up creating an error. My guards for moving
onto a non-existant tile weren't working with unexpectedly missing tiles. It only checked to see if the
player was off the board rather than if the tile simply didn't exist.